+++ packages/client/src/setupClient.ts
@@ -1,4 +1,4 @@
-import { append, create } from './utils/html';
+import { append, create } from './utils/dom';
 import { defineElements } from './elements/defineElements';
 import { CLIENT, InternalElements } from './constants';
 import { Options, setOptions } from './options';
